+++ src/bubble.h
@@ -9,4 +9,4 @@
 #include <stddef.h>
 
 
-void bubblesort(int* arr, size_t n);
+void bubblesort(int* arr, int n);
